Posterior predictive distributions for modeled probabilities

Get expected values of the posterior predictive distribution for the modeled probabilities (occupancy, detection, colonization, extinction, autologistic).

fitted_flocker(
  flocker_fit,
  components = c("occ", "det", "col", "ex", "auto", "Omega"),
  new_data = NULL,
  unit_level = FALSE,
  summarise = FALSE,
  CI = c(0.05, 0.95),
  draw_ids = NULL,
  response = TRUE,
  re_formula = NULL,
  allow_new_levels = FALSE,
  sample_new_levels = "uncertainty"
)

Arguments

flocker_fit

A flocker_fit object.

components

a character vector specifying one or more of "occ", "det", "col", "ex", "auto", and "Omega" for which to obtain fitted values.

new_data

Optional new data at which to evaluate occupancy predictions. New data can be passed as a flocker_data object produced by make_flocker_data or as a dataframe with one row per desired prediction. If `NULL` (the default) expected values are generated for the original data as formatted by make_flocker_data.

unit_level

Logical; defaults to FALSE. Relevant only when `new_data` is not a dataframe (i.e. it is `NULL` or a flocker_data object), and useful only for multiseason models with missing seasons. If FALSE, returns in the shape of the observation matrix/array with NAs for missing visits. If TRUE, returns in the shape of the first visit, and returns values for all units that are not part of a trailing block of never-visited units, including never-visited units that are part of series with subsequent visits.

summarise

if TRUE, return the expected value and upper and lower bound of the credible interval, otherwise return posterior draws.

CI

A vector of length 2 specifying the upper and lower bounds of the credible interval.

draw_ids

Vector of indices of the posterior draws to be used. If `NULL` (the default) all draws are used in their native order.

response

Should results be returned on the response or logit scale? Defaults to `TRUE`, i.e. response scale. However, the autologistic parameter is not interpretable as a probability and is always returned on the logit scale regardless of the value of `response`

re_formula

formula containing group-level effects to be considered in the prediction. If `NULL` (default), include all group-level effects; if NA, include no group-level effects.

allow_new_levels

allow new levels for random effect terms in `new_data`? Will error if set to `FALSE` and new levels are provided in `new_data`.

sample_new_levels

If new_data is provided and contains random effect levels not present in the original data, how should predictions be handled? Passed directly to `brms::prepare_predictions`, which see.

Value

A list of sets of expected values (one per component). If `new_data` is a dataframe, each element contains one row per row of `new_data`. Otherwise, returns in the shape of the observation matrix/array used to format the flocker_data (but see `unit_level` parameter for further details).

Details

The probabilities returned are conditional probabilities (e.g. detection conditional on occupancy, colonization conditional on previous non-occupancy, etc). These probabilities are not conditioned on the observed histories (e.g. the occupancy probability is not fixed to one at sites with a detection; it is estimated only based on the covariates).

Examples

# \donttest{
fitted_flocker(
  example_flocker_model_single, 
  summarise = TRUE
)
#> $linpred_occ
#> , , mean
#> 
#>             visit_1    visit_2    visit_3    visit_4
#> site_1   0.60096505 0.60096505 0.60096505 0.60096505
#> site_2   0.50105242 0.50105242 0.50105242 0.50105242
#> site_3   0.54987749 0.54987749 0.54987749 0.54987749
#> site_4   0.59048084 0.59048084 0.59048084 0.59048084
#> site_5   0.45610659 0.45610659 0.45610659 0.45610659
#> site_6   0.53478748 0.53478748 0.53478748 0.53478748
#> site_7   0.56750897 0.56750897 0.56750897 0.56750897
#> site_8   0.49521431 0.49521431 0.49521431 0.49521431
#> site_9   0.49696835 0.49696835 0.49696835 0.49696835
#> site_10  0.50265875 0.50265875 0.50265875 0.50265875
#> site_11  0.51501043 0.51501043 0.51501043 0.51501043
#> site_12  0.52592030 0.52592030 0.52592030 0.52592030
#> site_13  0.55912019 0.55912019 0.55912019 0.55912019
#> site_14  0.56787075 0.56787075 0.56787075 0.56787075
#> site_15  0.57028658 0.57028658 0.57028658 0.57028658
#> site_16  0.57952601 0.57952601 0.57952601 0.57952601
#> site_17  0.56452608 0.56452608 0.56452608 0.56452608
#> site_18  0.59321441 0.59321441 0.59321441 0.59321441
#> site_19  0.39512334 0.39512334 0.39512334 0.39512334
#> site_20  0.46106651 0.46106651 0.46106651 0.46106651
#> site_21  0.46253160 0.46253160 0.46253160 0.46253160
#> site_22  0.57616777 0.57616777 0.57616777 0.57616777
#> site_23  0.52793357 0.52793357 0.52793357 0.52793357
#> site_24  0.46868014 0.46868014 0.46868014 0.46868014
#> site_25  0.58842734 0.58842734 0.58842734 0.58842734
#> site_26  0.55086542 0.55086542 0.55086542 0.55086542
#> site_27  0.49437920 0.49437920 0.49437920 0.49437920
#> site_28  0.57854177 0.57854177 0.57854177 0.57854177
#> site_29  0.57786548 0.57786548 0.57786548 0.57786548
#> site_30  0.57544845 0.57544845 0.57544845 0.57544845
#> site_31  0.56869474 0.56869474 0.56869474 0.56869474
#> site_32  0.56024883 0.56024883 0.56024883 0.56024883
#> site_33  0.51017580 0.51017580 0.51017580 0.51017580
#> site_34  0.49375352 0.49375352 0.49375352 0.49375352
#> site_35  0.48974663 0.48974663 0.48974663 0.48974663
#> site_36  0.47754083 0.47754083 0.47754083 0.47754083
#> site_37  0.49975635 0.49975635 0.49975635 0.49975635
#> site_38  0.46702595 0.46702595 0.46702595 0.46702595
#> site_39  0.59643517 0.59643517 0.59643517 0.59643517
#> site_40  0.58754716 0.58754716 0.58754716 0.58754716
#> site_41  0.95113904 0.95113904 0.95113904 0.95113904
#> site_42  0.20300020 0.20300020 0.20300020 0.20300020
#> site_43  0.48239902 0.48239902 0.48239902 0.48239902
#> site_44  0.90537978 0.90537978 0.90537978 0.90537978
#> site_45  0.12352182 0.12352182 0.12352182 0.12352182
#> site_46  0.34649709 0.34649709 0.34649709 0.34649709
#> site_47  0.71236397 0.71236397 0.71236397 0.71236397
#> site_48  0.18860383 0.18860383 0.18860383 0.18860383
#> site_49  0.19273924 0.19273924 0.19273924 0.19273924
#> site_50  0.20730015 0.20730015 0.20730015 0.20730015
#> site_51  0.24659727 0.24659727 0.24659727 0.24659727
#> site_52  0.29409481 0.29409481 0.29409481 0.29409481
#> site_53  0.60057872 0.60057872 0.60057872 0.60057872
#> site_54  0.71683823 0.71683823 0.71683823 0.71683823
#> site_55  0.74553252 0.74553252 0.74553252 0.74553252
#> site_56  0.83463758 0.83463758 0.83463758 0.83463758
#> site_57  0.67399243 0.67399243 0.67399243 0.67399243
#> site_58  0.91895337 0.91895337 0.91895337 0.91895337
#> site_59  0.05893805 0.05893805 0.05893805 0.05893805
#> site_60  0.12980622 0.12980622 0.12980622 0.12980622
#> site_61  0.75677847 0.75677847 0.75677847 0.75677847
#> site_62  0.49918822 0.49918822 0.49918822 0.49918822
#> site_63  0.57925675 0.57925675 0.57925675 0.57925675
#> site_64  0.71577441 0.71577441 0.71577441 0.71577441
#> site_65  0.45416704 0.45416704 0.45416704 0.45416704
#> site_66  0.54684311 0.54684311 0.54684311 0.54684311
#> site_67  0.63146340 0.63146340 0.63146340 0.63146340
#> site_68  0.49275884 0.49275884 0.49275884 0.49275884
#> site_69  0.49465745 0.49465745 0.49465745 0.49465745
#> site_70  0.50101685 0.50101685 0.50101685 0.50101685
#> site_71  0.51619234 0.51619234 0.51619234 0.51619234
#> site_72  0.53184539 0.53184539 0.53184539 0.53184539
#> site_73  0.60464886 0.60464886 0.60464886 0.60464886
#> site_74  0.63269019 0.63269019 0.63269019 0.63269019
#> site_75  0.64100838 0.64100838 0.64100838 0.64100838
#> site_76  0.67439530 0.67439530 0.67439530 0.67439530
#> site_77  0.62155533 0.62155533 0.62155533 0.62155533
#> site_78  0.72628378 0.72628378 0.72628378 0.72628378
#> site_79  0.37106182 0.37106182 0.37106182 0.37106182
#> site_80  0.45893043 0.45893043 0.45893043 0.45893043
#> site_81  0.78693244 0.78693244 0.78693244 0.78693244
#> site_82  0.46308713 0.46308713 0.46308713 0.46308713
#> site_83  0.56405503 0.56405503 0.56405503 0.56405503
#> site_84  0.73606620 0.73606620 0.73606620 0.73606620
#> site_85  0.40868956 0.40868956 0.40868956 0.40868956
#> site_86  0.52274622 0.52274622 0.52274622 0.52274622
#> site_87  0.63064713 0.63064713 0.63064713 0.63064713
#> site_88  0.45518963 0.45518963 0.45518963 0.45518963
#> site_89  0.45751746 0.45751746 0.45751746 0.45751746
#> site_90  0.46534081 0.46534081 0.46534081 0.46534081
#> site_91  0.48416534 0.48416534 0.48416534 0.48416534
#> site_92  0.50378710 0.50378710 0.50378710 0.50378710
#> site_93  0.59650853 0.59650853 0.59650853 0.59650853
#> site_94  0.63220307 0.63220307 0.63220307 0.63220307
#> site_95  0.64273632 0.64273632 0.64273632 0.64273632
#> site_96  0.68469340 0.68469340 0.68469340 0.68469340
#> site_97  0.61805951 0.61805951 0.61805951 0.61805951
#> site_98  0.74906296 0.74906296 0.74906296 0.74906296
#> site_99  0.31229154 0.31229154 0.31229154 0.31229154
#> site_100 0.41434504 0.41434504 0.41434504 0.41434504
#> 
#> , , Q5
#> 
#>               visit_1      visit_2      visit_3      visit_4
#> site_1   2.529459e-01 2.529459e-01 2.529459e-01 2.529459e-01
#> site_2   3.110603e-01 3.110603e-01 3.110603e-01 3.110603e-01
#> site_3   3.879779e-01 3.879779e-01 3.879779e-01 3.879779e-01
#> site_4   3.108024e-01 3.108024e-01 3.108024e-01 3.108024e-01
#> site_5   2.637311e-01 2.637311e-01 2.637311e-01 2.637311e-01
#> site_6   3.574485e-01 3.574485e-01 3.574485e-01 3.574485e-01
#> site_7   3.832239e-01 3.832239e-01 3.832239e-01 3.832239e-01
#> site_8   3.050930e-01 3.050930e-01 3.050930e-01 3.050930e-01
#> site_9   3.068478e-01 3.068478e-01 3.068478e-01 3.068478e-01
#> site_10  3.127703e-01 3.127703e-01 3.127703e-01 3.127703e-01
#> site_11  3.271782e-01 3.271782e-01 3.271782e-01 3.271782e-01
#> site_12  3.424414e-01 3.424414e-01 3.424414e-01 3.424414e-01
#> site_13  3.955671e-01 3.955671e-01 3.955671e-01 3.955671e-01
#> site_14  3.822125e-01 3.822125e-01 3.822125e-01 3.822125e-01
#> site_15  3.753376e-01 3.753376e-01 3.753376e-01 3.753376e-01
#> site_16  3.472857e-01 3.472857e-01 3.472857e-01 3.472857e-01
#> site_17  3.881366e-01 3.881366e-01 3.881366e-01 3.881366e-01
#> site_18  3.003085e-01 3.003085e-01 3.003085e-01 3.003085e-01
#> site_19  7.142925e-02 7.142925e-02 7.142925e-02 7.142925e-02
#> site_20  2.702799e-01 2.702799e-01 2.702799e-01 2.702799e-01
#> site_21  1.173167e-01 1.173167e-01 1.173167e-01 1.173167e-01
#> site_22  3.901254e-01 3.901254e-01 3.901254e-01 3.901254e-01
#> site_23  4.512322e-01 4.512322e-01 4.512322e-01 4.512322e-01
#> site_24  1.531115e-01 1.531115e-01 1.531115e-01 1.531115e-01
#> site_25  3.365111e-01 3.365111e-01 3.365111e-01 3.365111e-01
#> site_26  4.454373e-01 4.454373e-01 4.454373e-01 4.454373e-01
#> site_27  2.889878e-01 2.889878e-01 2.889878e-01 2.889878e-01
#> site_28  3.824589e-01 3.824589e-01 3.824589e-01 3.824589e-01
#> site_29  3.847226e-01 3.847226e-01 3.847226e-01 3.847226e-01
#> site_30  3.923061e-01 3.923061e-01 3.923061e-01 3.923061e-01
#> site_31  4.104081e-01 4.104081e-01 4.104081e-01 4.104081e-01
#> site_32  4.290885e-01 4.290885e-01 4.290885e-01 4.290885e-01
#> site_33  3.698130e-01 3.698130e-01 3.698130e-01 3.698130e-01
#> site_34  2.858177e-01 2.858177e-01 2.858177e-01 2.858177e-01
#> site_35  2.656112e-01 2.656112e-01 2.656112e-01 2.656112e-01
#> site_36  2.060274e-01 2.060274e-01 2.060274e-01 2.060274e-01
#> site_37  3.163552e-01 3.163552e-01 3.163552e-01 3.163552e-01
#> site_38  1.428549e-01 1.428549e-01 1.428549e-01 1.428549e-01
#> site_39  2.390242e-01 2.390242e-01 2.390242e-01 2.390242e-01
#> site_40  3.421720e-01 3.421720e-01 3.421720e-01 3.421720e-01
#> site_41  8.369984e-01 8.369984e-01 8.369984e-01 8.369984e-01
#> site_42  2.801549e-02 2.801549e-02 2.801549e-02 2.801549e-02
#> site_43  2.678734e-01 2.678734e-01 2.678734e-01 2.678734e-01
#> site_44  7.268824e-01 7.268824e-01 7.268824e-01 7.268824e-01
#> site_45  4.845345e-03 4.845345e-03 4.845345e-03 4.845345e-03
#> site_46  1.418080e-01 1.418080e-01 1.418080e-01 1.418080e-01
#> site_47  4.618054e-01 4.618054e-01 4.618054e-01 4.618054e-01
#> site_48  2.205266e-02 2.205266e-02 2.205266e-02 2.205266e-02
#> site_49  2.367681e-02 2.367681e-02 2.367681e-02 2.367681e-02
#> site_50  2.996816e-02 2.996816e-02 2.996816e-02 2.996816e-02
#> site_51  5.177532e-02 5.177532e-02 5.177532e-02 5.177532e-02
#> site_52  8.867330e-02 8.867330e-02 8.867330e-02 8.867330e-02
#> site_53  3.738781e-01 3.738781e-01 3.738781e-01 3.738781e-01
#> site_54  4.659307e-01 4.659307e-01 4.659307e-01 4.659307e-01
#> site_55  4.938984e-01 4.938984e-01 4.938984e-01 4.938984e-01
#> site_56  6.009452e-01 6.009452e-01 6.009452e-01 6.009452e-01
#> site_57  4.287535e-01 4.287535e-01 4.287535e-01 4.287535e-01
#> site_58  7.557042e-01 7.557042e-01 7.557042e-01 7.557042e-01
#> site_59  9.981765e-05 9.981765e-05 9.981765e-05 9.981765e-05
#> site_60  5.886503e-03 5.886503e-03 5.886503e-03 5.886503e-03
#> site_61  4.623995e-01 4.623995e-01 4.623995e-01 4.623995e-01
#> site_62  2.757362e-01 2.757362e-01 2.757362e-01 2.757362e-01
#> site_63  4.249628e-01 4.249628e-01 4.249628e-01 4.249628e-01
#> site_64  4.643492e-01 4.643492e-01 4.643492e-01 4.643492e-01
#> site_65  2.039495e-01 2.039495e-01 2.039495e-01 2.039495e-01
#> site_66  3.637124e-01 3.637124e-01 3.637124e-01 3.637124e-01
#> site_67  4.664820e-01 4.664820e-01 4.664820e-01 4.664820e-01
#> site_68  2.647663e-01 2.647663e-01 2.647663e-01 2.647663e-01
#> site_69  2.679818e-01 2.679818e-01 2.679818e-01 2.679818e-01
#> site_70  2.788980e-01 2.788980e-01 2.788980e-01 2.788980e-01
#> site_71  3.058288e-01 3.058288e-01 3.058288e-01 3.058288e-01
#> site_72  3.348381e-01 3.348381e-01 3.348381e-01 3.348381e-01
#> site_73  4.594864e-01 4.594864e-01 4.594864e-01 4.594864e-01
#> site_74  4.664504e-01 4.664504e-01 4.664504e-01 4.664504e-01
#> site_75  4.662420e-01 4.662420e-01 4.662420e-01 4.662420e-01
#> site_76  4.654564e-01 4.654564e-01 4.654564e-01 4.654564e-01
#> site_77  4.667452e-01 4.667452e-01 4.667452e-01 4.667452e-01
#> site_78  4.639764e-01 4.639764e-01 4.639764e-01 4.639764e-01
#> site_79  9.381481e-02 9.381481e-02 9.381481e-02 9.381481e-02
#> site_80  2.109752e-01 2.109752e-01 2.109752e-01 2.109752e-01
#> site_81  5.970610e-01 5.970610e-01 5.970610e-01 5.970610e-01
#> site_82  2.781566e-01 2.781566e-01 2.781566e-01 2.781566e-01
#> site_83  4.139642e-01 4.139642e-01 4.139642e-01 4.139642e-01
#> site_84  5.526875e-01 5.526875e-01 5.526875e-01 5.526875e-01
#> site_85  1.816135e-01 1.816135e-01 1.816135e-01 1.816135e-01
#> site_86  3.696546e-01 3.696546e-01 3.696546e-01 3.696546e-01
#> site_87  4.730126e-01 4.730126e-01 4.730126e-01 4.730126e-01
#> site_88  2.630930e-01 2.630930e-01 2.630930e-01 2.630930e-01
#> site_89  2.675006e-01 2.675006e-01 2.675006e-01 2.675006e-01
#> site_90  2.825113e-01 2.825113e-01 2.825113e-01 2.825113e-01
#> site_91  3.166497e-01 3.166497e-01 3.166497e-01 3.166497e-01
#> site_92  3.465845e-01 3.465845e-01 3.465845e-01 3.465845e-01
#> site_93  4.445354e-01 4.445354e-01 4.445354e-01 4.445354e-01
#> site_94  4.741903e-01 4.741903e-01 4.741903e-01 4.741903e-01
#> site_95  4.822577e-01 4.822577e-01 4.822577e-01 4.822577e-01
#> site_96  5.157569e-01 5.157569e-01 5.157569e-01 5.157569e-01
#> site_97  4.636058e-01 4.636058e-01 4.636058e-01 4.636058e-01
#> site_98  5.631479e-01 5.631479e-01 5.631479e-01 5.631479e-01
#> site_99  6.327459e-02 6.327459e-02 6.327459e-02 6.327459e-02
#> site_100 1.907875e-01 1.907875e-01 1.907875e-01 1.907875e-01
#> 
#> , , Q95
#> 
#>            visit_1   visit_2   visit_3   visit_4
#> site_1   0.9320204 0.9320204 0.9320204 0.9320204
#> site_2   0.6794074 0.6794074 0.6794074 0.6794074
#> site_3   0.8095561 0.8095561 0.8095561 0.8095561
#> site_4   0.9102895 0.9102895 0.9102895 0.9102895
#> site_5   0.6690977 0.6690977 0.6690977 0.6690977
#> site_6   0.7819317 0.7819317 0.7819317 0.7819317
#> site_7   0.8605867 0.8605867 0.8605867 0.8605867
#> site_8   0.6587020 0.6587020 0.6587020 0.6587020
#> site_9   0.6649773 0.6649773 0.6649773 0.6649773
#> site_10  0.6850066 0.6850066 0.6850066 0.6850066
#> site_11  0.7262449 0.7262449 0.7262449 0.7262449
#> site_12  0.7590139 0.7590139 0.7590139 0.7590139
#> site_13  0.8381731 0.8381731 0.8381731 0.8381731
#> site_14  0.8614824 0.8614824 0.8614824 0.8614824
#> site_15  0.8673291 0.8673291 0.8673291 0.8673291
#> site_16  0.8879144 0.8879144 0.8879144 0.8879144
#> site_17  0.8529879 0.8529879 0.8529879 0.8529879
#> site_18  0.9158131 0.9158131 0.9158131 0.9158131
#> site_19  0.7703219 0.7703219 0.7703219 0.7703219
#> site_20  0.6634769 0.6634769 0.6634769 0.6634769
#> site_21  0.7764400 0.7764400 0.7764400 0.7764400
#> site_22  0.7900526 0.7900526 0.7900526 0.7900526
#> site_23  0.6531165 0.6531165 0.6531165 0.6531165
#> site_24  0.7241923 0.7241923 0.7241923 0.7241923
#> site_25  0.8585616 0.8585616 0.8585616 0.8585616
#> site_26  0.6723555 0.6723555 0.6723555 0.6723555
#> site_27  0.6591839 0.6591839 0.6591839 0.6591839
#> site_28  0.8017801 0.8017801 0.8017801 0.8017801
#> site_29  0.7984482 0.7984482 0.7984482 0.7984482
#> site_30  0.7864835 0.7864835 0.7864835 0.7864835
#> site_31  0.7527299 0.7527299 0.7527299 0.7527299
#> site_32  0.7101534 0.7101534 0.7101534 0.7101534
#> site_33  0.6492689 0.6492689 0.6492689 0.6492689
#> site_34  0.6600364 0.6600364 0.6600364 0.6600364
#> site_35  0.6658383 0.6658383 0.6658383 0.6658383
#> site_36  0.6895986 0.6895986 0.6895986 0.6895986
#> site_37  0.6523258 0.6523258 0.6523258 0.6523258
#> site_38  0.7372773 0.7372773 0.7372773 0.7372773
#> site_39  0.9220681 0.9220681 0.9220681 0.9220681
#> site_40  0.8524508 0.8524508 0.8524508 0.8524508
#> site_41  0.9994623 0.9994623 0.9994623 0.9994623
#> site_42  0.5122035 0.5122035 0.5122035 0.5122035
#> site_43  0.7770863 0.7770863 0.7770863 0.7770863
#> site_44  0.9955984 0.9955984 0.9955984 0.9955984
#> site_45  0.4005576 0.4005576 0.4005576 0.4005576
#> site_46  0.6670404 0.6670404 0.6670404 0.6670404
#> site_47  0.8976755 0.8976755 0.8976755 0.8976755
#> site_48  0.4928141 0.4928141 0.4928141 0.4928141
#> site_49  0.4984507 0.4984507 0.4984507 0.4984507
#> site_50  0.5178671 0.5178671 0.5178671 0.5178671
#> site_51  0.5668060 0.5668060 0.5668060 0.5668060
#> site_52  0.6188264 0.6188264 0.6188264 0.6188264
#> site_53  0.8450100 0.8450100 0.8450100 0.8450100
#> site_54  0.9012196 0.9012196 0.9012196 0.9012196
#> site_55  0.9228126 0.9228126 0.9228126 0.9228126
#> site_56  0.9752298 0.9752298 0.9752298 0.9752298
#> site_57  0.8762110 0.8762110 0.8762110 0.8762110
#> site_58  0.9972996 0.9972996 0.9972996 0.9972996
#> site_59  0.2687623 0.2687623 0.2687623 0.2687623
#> site_60  0.4084579 0.4084579 0.4084579 0.4084579
#> site_61  0.9510832 0.9510832 0.9510832 0.9510832
#> site_62  0.8419114 0.8419114 0.8419114 0.8419114
#> site_63  0.8832979 0.8832979 0.8832979 0.8832979
#> site_64  0.9351751 0.9351751 0.9351751 0.9351751
#> site_65  0.8105593 0.8105593 0.8105593 0.8105593
#> site_66  0.8683829 0.8683829 0.8683829 0.8683829
#> site_67  0.9034112 0.9034112 0.9034112 0.9034112
#> site_68  0.8378474 0.8378474 0.8378474 0.8378474
#> site_69  0.8390609 0.8390609 0.8390609 0.8390609
#> site_70  0.8430440 0.8430440 0.8430440 0.8430440
#> site_71  0.8520622 0.8520622 0.8520622 0.8520622
#> site_72  0.8606943 0.8606943 0.8606943 0.8606943
#> site_73  0.8936027 0.8936027 0.8936027 0.8936027
#> site_74  0.9038370 0.9038370 0.9038370 0.9038370
#> site_75  0.9066751 0.9066751 0.9066751 0.9066751
#> site_76  0.9173083 0.9173083 0.9173083 0.9173083
#> site_77  0.8999023 0.8999023 0.8999023 0.8999023
#> site_78  0.9394700 0.9394700 0.9394700 0.9394700
#> site_79  0.8061009 0.8061009 0.8061009 0.8061009
#> site_80  0.8142126 0.8142126 0.8142126 0.8142126
#> site_81  0.9554809 0.9554809 0.9554809 0.9554809
#> site_82  0.7719124 0.7719124 0.7719124 0.7719124
#> site_83  0.8372616 0.8372616 0.8372616 0.8372616
#> site_84  0.9237000 0.9237000 0.9237000 0.9237000
#> site_85  0.7247656 0.7247656 0.7247656 0.7247656
#> site_86  0.8130531 0.8130531 0.8130531 0.8130531
#> site_87  0.8717248 0.8717248 0.8717248 0.8717248
#> site_88  0.7657368 0.7657368 0.7657368 0.7657368
#> site_89  0.7675780 0.7675780 0.7675780 0.7675780
#> site_90  0.7736386 0.7736386 0.7736386 0.7736386
#> site_91  0.7874758 0.7874758 0.7874758 0.7874758
#> site_92  0.8009045 0.8009045 0.8009045 0.8009045
#> site_93  0.8546120 0.8546120 0.8546120 0.8546120
#> site_94  0.8724826 0.8724826 0.8724826 0.8724826
#> site_95  0.8775690 0.8775690 0.8775690 0.8775690
#> site_96  0.8972056 0.8972056 0.8972056 0.8972056
#> site_97  0.8655285 0.8655285 0.8655285 0.8655285
#> site_98  0.9327258 0.9327258 0.9327258 0.9327258
#> site_99  0.6402168 0.6402168 0.6402168 0.6402168
#> site_100 0.7302212 0.7302212 0.7302212 0.7302212
#> 
#> 
#> $linpred_det
#> , , mean
#> 
#>             visit_1    visit_2    visit_3    visit_4
#> site_1   0.09067892 0.47540359 0.37195810 0.11536463
#> site_2   0.27239331 0.69791653 0.49870285 0.25398023
#> site_3   0.55169667 0.75285138 0.18029473 0.57195343
#> site_4   0.31469312 0.23418009 0.09398465 0.16868052
#> site_5   0.59099518 0.74861959 0.57111711 0.79935320
#> site_6   0.37585724 0.67503968 0.25360258 0.22388280
#> site_7   0.77955634 0.22994519 0.36678828 0.44210378
#> site_8   0.46288451 0.67747385 0.64985081 0.52309045
#> site_9   0.58250221 0.56026777 0.84660360 0.40440454
#> site_10  0.33712029 0.39744195 0.61127437 0.49833713
#> site_11  0.44343670 0.32159615 0.51275148 0.35227695
#> site_12  0.19419962 0.40414731 0.67871029 0.36315895
#> site_13  0.48643236 0.12678895 0.33768584 0.46670253
#> site_14  0.36522760 0.32324173 0.19679019 0.17182347
#> site_15  0.14643404 0.31173645 0.15366800 0.20700540
#> site_16  0.19914207 0.58346383 0.15218669 0.28078343
#> site_17  0.33651367 0.17092814 0.30578361 0.61275685
#> site_18  0.23817552 0.17641465 0.12711729 0.05000732
#> site_19  0.89160342 0.54980771 0.86847842 0.92084465
#> site_20  0.33806787 0.32407788 0.52646897 0.34680821
#> site_21  0.43988013 0.65272257 0.61089485 0.47215302
#> site_22  0.44144978 0.63369899 0.55240375 0.42947945
#> site_23  0.61180734 0.70845587 0.43195007 0.61968829
#> site_24  0.57445705 0.53866726 0.41784568 0.49919329
#> site_25  0.56410161 0.63430723 0.55590533 0.66496379
#> site_26  0.52823005 0.64813378 0.46411961 0.44399706
#> site_27  0.75423064 0.49957060 0.56442318 0.59282247
#> site_28  0.53400858 0.62082852 0.60885913 0.55850118
#> site_29  0.58273491 0.57407323 0.72823815 0.50936901
#> site_30  0.48036407 0.51031500 0.59734241 0.55327705
#> site_31  0.53963964 0.48274260 0.56703917 0.49861519
#> site_32  0.41035247 0.53205399 0.64201549 0.51377185
#> site_33  0.59831183 0.39405025 0.54097443 0.59103225
#> site_34  0.56430506 0.54715933 0.47803100 0.45813690
#> site_35  0.43913764 0.54568332 0.44640258 0.48975589
#> site_36  0.50118581 0.66652504 0.46425665 0.54533471
#> site_37  0.54802984 0.45069575 0.53456363 0.65622905
#> site_38  0.54440413 0.50936435 0.46784504 0.32495533
#> site_39  0.69382891 0.48916539 0.66756494 0.73956807
#> site_40  0.44857820 0.44054952 0.54010166 0.45348866
#> site_41  0.57700492 0.18550493 0.25432213 0.51818627
#> site_42  0.34834944 0.11792250 0.19883457 0.36651865
#> site_43  0.14519356 0.07240433 0.39463455 0.13763689
#> site_44  0.26108250 0.32407858 0.55985334 0.39895759
#> site_45  0.18385820 0.11663617 0.19287685 0.09320466
#> site_46  0.23520662 0.10937574 0.32728048 0.35916135
#> site_47  0.05586103 0.31068163 0.21393389 0.17841775
#> site_48  0.22019539 0.12866552 0.13949334 0.19139263
#> site_49  0.16525206 0.17437444 0.05753023 0.25156919
#> site_50  0.29192334 0.25110974 0.15086980 0.19797262
#> site_51  0.21575343 0.29141270 0.18370479 0.26915457
#> site_52  0.40825748 0.22747583 0.11301998 0.25141913
#> site_53  0.16426655 0.48230832 0.23464086 0.17228979
#> site_54  0.21463085 0.23821259 0.34820394 0.38387558
#> site_55  0.42688276 0.24471801 0.41297081 0.33403901
#> site_56  0.34340526 0.11678591 0.41284187 0.26801077
#> site_57  0.23179550 0.38876773 0.25084564 0.11279135
#> site_58  0.32838132 0.39526251 0.47682643 0.73958621
#> site_59  0.07652920 0.27283999 0.09464380 0.05077897
#> site_60  0.33022923 0.34160377 0.21102509 0.32332719
#> site_61  0.29387388 0.39680095 0.37024737 0.30551535
#> site_62  0.33107965 0.41893508 0.37831067 0.32623425
#> site_63  0.39416382 0.45477065 0.31474842 0.39844290
#> site_64  0.35546889 0.33845857 0.29143180 0.32168455
#> site_65  0.39250572 0.42756421 0.38870658 0.44468379
#> site_66  0.35913912 0.41946030 0.33234498 0.32431633
#> site_67  0.48715019 0.33385909 0.36243270 0.37652589
#> site_68  0.37120957 0.41311955 0.40686363 0.38231375
#> site_69  0.39344982 0.38927875 0.47939810 0.36010098
#> site_70  0.34670582 0.35934195 0.39950648 0.37837448
#> site_71  0.36945392 0.34502998 0.38203907 0.35166044
#> site_72  0.31334260 0.36335209 0.41821133 0.35540054
#> site_73  0.38325621 0.29686786 0.35561220 0.37950427
#> site_74  0.36219675 0.35421074 0.32500494 0.31724572
#> site_75  0.30888281 0.35231793 0.31161782 0.32848680
#> site_76  0.32834113 0.41455996 0.31380318 0.34733350
#> site_77  0.35625746 0.31594876 0.35021328 0.41411954
#> site_78  0.33959373 0.32435958 0.30804742 0.25898766
#> site_79  0.47867139 0.37827568 0.46399725 0.50647022
#> site_80  0.34132215 0.33799339 0.38053779 0.34336200
#> site_81  0.58705359 0.17509401 0.23746967 0.52210953
#> site_82  0.49965886 0.19210257 0.30109901 0.52270667
#> site_83  0.22528428 0.12031256 0.53669483 0.21502930
#> site_84  0.29739065 0.36585300 0.60755924 0.44665142
#> site_85  0.27649072 0.18726079 0.28865287 0.15586030
#> site_86  0.34891221 0.17920175 0.46687676 0.50554091
#> site_87  0.08183561 0.41985433 0.30188783 0.25675439
#> site_88  0.32955537 0.20644314 0.22102307 0.29071578
#> site_89  0.25570489 0.26795280 0.10291841 0.37197799
#> site_90  0.42599048 0.37152289 0.23671116 0.30002449
#> site_91  0.32429202 0.42474488 0.28127513 0.39547724
#> site_92  0.56631677 0.33965318 0.18513321 0.37137211
#> site_93  0.24607854 0.61372947 0.33744366 0.25665751
#> site_94  0.30227110 0.33173975 0.46211124 0.50102663
#> site_95  0.54037066 0.33590891 0.52637027 0.44151709
#> site_96  0.42869423 0.16148275 0.50238805 0.34319768
#> site_97  0.32820822 0.51221841 0.35204778 0.17288889
#> site_98  0.35716681 0.42942100 0.51568912 0.75723427
#> site_99  0.12237223 0.38653120 0.14445672 0.08872670
#> site_100 0.47775978 0.49323726 0.31393971 0.46833209
#> 
#> , , Q5
#> 
#>               visit_1     visit_2      visit_3      visit_4
#> site_1   0.0091478464 0.044385868 0.0307529643 0.0112628350
#> site_2   0.0797030441 0.443001246 0.3301396042 0.0656832951
#> site_3   0.2805442113 0.408862852 0.0519265935 0.2884861230
#> site_4   0.0483088552 0.037852654 0.0177162949 0.0293504375
#> site_5   0.4125455693 0.508852511 0.3792205266 0.5427743737
#> site_6   0.2095392348 0.372940423 0.0925489471 0.0709339927
#> site_7   0.4101866101 0.094294866 0.1362538954 0.1609229046
#> site_8   0.2744347004 0.438804857 0.4233117646 0.3577569901
#> site_9   0.3877591492 0.377810638 0.5707443695 0.2009528490
#> site_10  0.1388761591 0.198933014 0.3926972849 0.3297781245
#> site_11  0.2720118068 0.133470024 0.3311777839 0.1627135570
#> site_12  0.0421245716 0.234450983 0.3914342544 0.1893991731
#> site_13  0.2166674225 0.023822170 0.1591031331 0.2082404990
#> site_14  0.1343364484 0.121694815 0.0757310184 0.0598030010
#> site_15  0.0435047879 0.109712845 0.0489119007 0.0795731182
#> site_16  0.0553344795 0.165763980 0.0444164182 0.0724042686
#> site_17  0.1382887178 0.056224514 0.1283055851 0.2594797908
#> site_18  0.0329830042 0.026159588 0.0201044467 0.0031545603
#> site_19  0.6750896438 0.204408614 0.6478068539 0.7220599416
#> site_20  0.1028308273 0.091050936 0.3210082681 0.1106890031
#> site_21  0.2233438226 0.371302338 0.4124211245 0.2539440516
#> site_22  0.3181206266 0.486005013 0.4738025940 0.3019007758
#> site_23  0.5113495548 0.393215258 0.3343568610 0.5017745230
#> site_24  0.4184361277 0.370523020 0.2406928496 0.3234205481
#> site_25  0.4769247348 0.484671690 0.4647278360 0.4469984910
#> site_26  0.4509500016 0.467395933 0.3669462510 0.3412526028
#> site_27  0.3374648934 0.387062062 0.4642890998 0.5016633685
#> site_28  0.4452990738 0.501612638 0.5157961238 0.4807716581
#> site_29  0.5177749361 0.504607167 0.3762171570 0.4115325063
#> site_30  0.3741263427 0.414794879 0.5249219930 0.4756425112
#> site_31  0.4605436367 0.382345123 0.5006086208 0.4035103200
#> site_32  0.2914733619 0.454469820 0.4754132617 0.4288894558
#> site_33  0.5222962420 0.288733725 0.4497937458 0.5132907830
#> site_34  0.4634710556 0.442036520 0.3627663098 0.3416518557
#> site_35  0.3177017446 0.435633037 0.3251055421 0.3709620625
#> site_36  0.3614709616 0.406817330 0.3204407668 0.4152141196
#> site_37  0.4491976452 0.339735214 0.4329027385 0.4435995541
#> site_38  0.3678363097 0.325098206 0.2799967433 0.1545375937
#> site_39  0.4079485714 0.281076551 0.4383877229 0.3636716584
#> site_40  0.3029921982 0.290868205 0.4426515034 0.3105149040
#> site_41  0.3036891866 0.026577172 0.0506237125 0.2388982994
#> site_42  0.0567502383 0.008280418 0.0199304387 0.0633472545
#> site_43  0.0294361413 0.008100721 0.1586485626 0.0269012762
#> site_44  0.1059608942 0.155268569 0.3755857733 0.2237588695
#> site_45  0.0097003521 0.004479889 0.0105493292 0.0030608536
#> site_46  0.0460750471 0.012728158 0.0824866037 0.0983813113
#> site_47  0.0060032035 0.168898973 0.0945259532 0.0707737805
#> site_48  0.0219762977 0.008847481 0.0101299022 0.0172455107
#> site_49  0.0137775556 0.015080501 0.0020676210 0.0285310681
#> site_50  0.0405790932 0.030673075 0.0128076018 0.0202307042
#> site_51  0.0281028381 0.047985753 0.0214064092 0.0414552892
#> site_52  0.1108648593 0.036839096 0.0113751908 0.0437786709
#> site_53  0.0471613982 0.280698474 0.0839901616 0.0510019571
#> site_54  0.0961538800 0.113190208 0.2037147080 0.2371139876
#> site_55  0.2980709224 0.127711042 0.2833312294 0.2051954331
#> site_56  0.2476388158 0.039355724 0.3203915185 0.1735136255
#> site_57  0.0973651876 0.221011434 0.1102728297 0.0293305967
#> site_58  0.1392878043 0.196009730 0.2777151935 0.5187966451
#> site_59  0.0005508145 0.005947404 0.0007897679 0.0002690798
#> site_60  0.0319234609 0.034465502 0.0132043910 0.0304637861
#> site_61  0.1645882722 0.260733884 0.2444754117 0.1777137527
#> site_62  0.2543019723 0.291998539 0.2838460870 0.2470091278
#> site_63  0.2933220673 0.303566408 0.2437895413 0.2946728496
#> site_64  0.2593750634 0.242307961 0.1924493002 0.2255765364
#> site_65  0.2739959581 0.280983637 0.2732408765 0.2844580644
#> site_66  0.2806966766 0.300045945 0.2689568544 0.2567303035
#> site_67  0.3057081443 0.266955390 0.2807509180 0.2874845874
#> site_68  0.2806498840 0.289056637 0.2878035079 0.2828881276
#> site_69  0.2856434746 0.284808051 0.2989721589 0.2785655732
#> site_70  0.2701752747 0.278610646 0.2886020447 0.2843581400
#> site_71  0.2826131808 0.272654140 0.2866978256 0.2768014962
#> site_72  0.2365770057 0.281350104 0.2990656439 0.2787729015
#> site_73  0.2911058034 0.214521700 0.2803725186 0.2899354930
#> site_74  0.2805020769 0.276675634 0.2573397953 0.2473239905
#> site_75  0.2356714655 0.274858934 0.2391633769 0.2609550706
#> site_76  0.2553775502 0.300358541 0.2377733380 0.2687307206
#> site_77  0.2788689782 0.246871647 0.2759488167 0.3015813985
#> site_78  0.2367965882 0.221840877 0.2055315715 0.1337688727
#> site_79  0.2646616160 0.230731206 0.2615049063 0.2710148992
#> site_80  0.2536810719 0.250771325 0.2729985235 0.2554784990
#> site_81  0.5010671767 0.054464504 0.1126990782 0.4504878989
#> site_82  0.3418053393 0.062288230 0.1495410813 0.3664982670
#> site_83  0.1006726600 0.017248138 0.4099793712 0.0901569354
#> site_84  0.1976655561 0.285670053 0.5089972151 0.3739606570
#> site_85  0.1130856790 0.055330182 0.1225280915 0.0332780291
#> site_86  0.2178803805 0.053758688 0.3337681697 0.3744059774
#> site_87  0.0054851069 0.331292281 0.1992958611 0.1416146524
#> site_88  0.1737253955 0.073797357 0.0835515025 0.1379436836
#> site_89  0.1095828120 0.119378823 0.0097475789 0.2176751395
#> site_90  0.2693113104 0.220229596 0.0967947819 0.1492475880
#> site_91  0.1784269351 0.275924079 0.1377772002 0.2486587904
#> site_92  0.4118368983 0.201019714 0.0579982951 0.2343041500
#> site_93  0.1256787109 0.474423288 0.2317566093 0.1378680171
#> site_94  0.2000519634 0.238574959 0.3717380292 0.4077931957
#> site_95  0.4365716032 0.247877661 0.4269645508 0.3540090944
#> site_96  0.3488679617 0.046035573 0.4189554239 0.2629128770
#> site_97  0.2289192968 0.410504528 0.2576906499 0.0523913928
#> site_98  0.2733897821 0.353529540 0.4456162716 0.6380630986
#> site_99  0.0149410192 0.165650316 0.0247411355 0.0053939882
#> site_100 0.2964667586 0.312137351 0.1453204829 0.2871376966
#> 
#> , , Q95
#> 
#>            visit_1   visit_2   visit_3   visit_4
#> site_1   0.2107678 0.8696771 0.7500323 0.2460456
#> site_2   0.3951888 0.8608784 0.6676430 0.3702388
#> site_3   0.7336955 0.9664497 0.2905315 0.7680842
#> site_4   0.5546557 0.4047794 0.2188728 0.3228077
#> site_5   0.7959336 0.9002894 0.7812396 0.9392798
#> site_6   0.5197469 0.8754841 0.3522614 0.3168994
#> site_7   0.9901143 0.3752418 0.5145360 0.6304242
#> site_8   0.6407716 0.8372110 0.8098425 0.6974658
#> site_9   0.7447044 0.7267417 0.9861894 0.5747397
#> site_10  0.4824568 0.5595258 0.7692342 0.6650656
#> site_11  0.6014933 0.4468951 0.6732724 0.4890896
#> site_12  0.2910159 0.5552669 0.8621420 0.5044856
#> site_13  0.6585491 0.2492458 0.4768765 0.6207872
#> site_14  0.5135980 0.4748271 0.3338118 0.2997796
#> site_15  0.2705750 0.4668309 0.2784551 0.3506820
#> site_16  0.3531337 0.8872964 0.2878023 0.4464981
#> site_17  0.4829946 0.2930419 0.4534871 0.8757000
#> site_18  0.4105022 0.3345788 0.2618073 0.1541842
#> site_19  0.9856486 0.8655256 0.9789372 0.9950144
#> site_20  0.5411407 0.5232964 0.7397866 0.5519848
#> site_21  0.7550260 0.8984765 0.8510513 0.7441342
#> site_22  0.6979988 0.7217049 0.6315969 0.7016196
#> site_23  0.7110502 0.8601604 0.6924446 0.7237390
#> site_24  0.7580199 0.7031211 0.7374080 0.7110881
#> site_25  0.6366537 0.7323458 0.6354945 0.7733337
#> site_26  0.6463476 0.7553900 0.6841244 0.6919191
#> site_27  0.9343939 0.6806842 0.6649545 0.7133702
#> site_28  0.6477307 0.7007594 0.6834272 0.6285635
#> site_29  0.6459897 0.6337892 0.8586952 0.6653566
#> site_30  0.6815417 0.6640111 0.6653102 0.6306256
#> site_31  0.6406656 0.6785940 0.6256077 0.6696991
#> site_32  0.7026631 0.6450266 0.7392823 0.6583049
#> site_33  0.7012646 0.6984559 0.6466137 0.6890896
#> site_34  0.6658002 0.6546731 0.6907275 0.6974449
#> site_35  0.7049417 0.6589732 0.7033735 0.6886040
#> site_36  0.6944399 0.8565044 0.7105357 0.6786252
#> site_37  0.6485851 0.6956820 0.6562829 0.8081378
#> site_38  0.7213476 0.7088459 0.7307832 0.7322370
#> site_39  0.8390510 0.7108684 0.8105107 0.8875916
#> site_40  0.7004751 0.7031966 0.6465687 0.6986623
#> site_41  0.8025128 0.3772411 0.4367973 0.7461771
#> site_42  0.6019860 0.2382156 0.3587240 0.6273888
#> site_43  0.2781796 0.2126414 0.5567682 0.2723734
#> site_44  0.4015431 0.4474210 0.7101574 0.5284486
#> site_45  0.3861094 0.2441650 0.4046082 0.2114335
#> site_46  0.3638348 0.2423093 0.4963377 0.5407610
#> site_47  0.2024509 0.4397964 0.3436835 0.3177382
#> site_48  0.4061369 0.2481154 0.2586381 0.3528436
#> site_49  0.2998807 0.3178021 0.1757783 0.4583701
#> site_50  0.5149036 0.4485560 0.2682899 0.3548198
#> site_51  0.3671026 0.4912978 0.3099983 0.4562564
#> site_52  0.6310543 0.3667692 0.2421122 0.4050654
#> site_53  0.2961917 0.6507867 0.3466166 0.3021564
#> site_54  0.3446607 0.3607617 0.4860874 0.5281549
#> site_55  0.5723000 0.3683444 0.5571743 0.4655766
#> site_56  0.4670277 0.2803745 0.5383905 0.3959427
#> site_57  0.3518892 0.5380417 0.3690542 0.2596067
#> site_58  0.4490823 0.5336675 0.6326283 0.8826911
#> site_59  0.2084111 0.6692642 0.2548161 0.1470947
#> site_60  0.6352999 0.6517439 0.4342793 0.6251110
#> site_61  0.4626303 0.5776122 0.5347313 0.4746203
#> site_62  0.4285409 0.5910677 0.5184745 0.4189454
#> site_63  0.4954131 0.6025043 0.3661982 0.5026968
#> site_64  0.4794183 0.4636330 0.4197250 0.4487862
#> site_65  0.5746758 0.6357581 0.5678198 0.6642695
#> site_66  0.4521659 0.5610228 0.4016304 0.3865013
#> site_67  0.6694198 0.4065403 0.4420744 0.4591531
#> site_68  0.5095919 0.5851540 0.5741613 0.5300675
#> site_69  0.5490025 0.5414769 0.6952890 0.4875133
#> site_70  0.4578350 0.4819312 0.5556431 0.5173865
#> site_71  0.4909959 0.4449599 0.5141433 0.4576078
#> site_72  0.3803669 0.4695695 0.5685455 0.4547089
#> site_73  0.4650365 0.3621230 0.4238357 0.4586169
#> site_74  0.4423909 0.4325908 0.3957988 0.3857162
#> site_75  0.3822814 0.4343622 0.3839629 0.4043070
#> site_76  0.4210465 0.5314246 0.4076517 0.4453507
#> site_77  0.4296866 0.3787814 0.4222452 0.5114181
#> site_78  0.4744398 0.4609451 0.4457076 0.3927958
#> site_79  0.7717550 0.6101980 0.7507435 0.8096264
#> site_80  0.4747888 0.4681750 0.5496294 0.4788211
#> site_81  0.7340756 0.2874741 0.3441060 0.6308366
#> site_82  0.6277976 0.3208863 0.4202979 0.6570948
#> site_83  0.3396799 0.2173779 0.6493486 0.3298053
#> site_84  0.3872561 0.4298430 0.7506094 0.4831191
#> site_85  0.4106386 0.3238938 0.4205649 0.2849895
#> site_86  0.4441429 0.2977060 0.5685149 0.6200515
#> site_87  0.1623989 0.4791720 0.3952441 0.3610123
#> site_88  0.4428073 0.3375678 0.3524123 0.4136413
#> site_89  0.3842339 0.3947259 0.1997772 0.4716506
#> site_90  0.5257793 0.4698500 0.3658178 0.4190753
#> site_91  0.4337642 0.5191253 0.4011159 0.4847698
#> site_92  0.7089592 0.4411461 0.3071418 0.4625895
#> site_93  0.3543039 0.7606747 0.4231654 0.3630752
#> site_94  0.3954068 0.4156858 0.5332313 0.5858318
#> site_95  0.6454139 0.4176610 0.6219716 0.5016946
#> site_96  0.4752717 0.2708449 0.5801707 0.4189764
#> site_97  0.4143423 0.6043924 0.4313741 0.2820473
#> site_98  0.4249190 0.4683093 0.6068389 0.9300400
#> site_99  0.2515784 0.5165516 0.2849971 0.1939983
#> site_100 0.6123718 0.6329178 0.4390977 0.5995665
#> 
#> 
# }