Posterior predictive distributions for modeled probabilities

Get expected values of the posterior predictive distribution for the modeled probabilities (occupancy, detection, colonization, extinction, autologistic).

fitted_flocker(
  flocker_fit,
  components = c("occ", "det", "col", "ex", "auto", "Omega"),
  new_data = NULL,
  unit_level = FALSE,
  summarise = FALSE,
  CI = c(0.05, 0.95),
  draw_ids = NULL,
  response = TRUE,
  re_formula = NULL,
  allow_new_levels = FALSE,
  sample_new_levels = "uncertainty"
)

Arguments

flocker_fit

A flocker_fit object.

components

a character vector specifying one or more of "occ", "det", "col", "ex", "auto", and "Omega" for which to obtain fitted values.

new_data

Optional new data at which to evaluate occupancy predictions. New data can be passed as a flocker_data object produced by make_flocker_data or as a dataframe with one row per desired prediction. If `NULL` (the default) expected values are generated for the original data as formatted by make_flocker_data.

unit_level

Logical; defaults to FALSE. Relevant only when `new_data` is not a dataframe (i.e. it is `NULL` or a flocker_data object), and useful only for multiseason models with missing seasons. If FALSE, returns in the shape of the observation matrix/array with NAs for missing visits. If TRUE, returns in the shape of the first visit, and returns values for all units that are not part of a trailing block of never-visited units, including never-visited units that are part of series with subsequent visits.

summarise

if TRUE, return the expected value and upper and lower bound of the credible interval, otherwise return posterior draws.

CI

A vector of length 2 specifying the upper and lower bounds of the credible interval.

draw_ids

Vector of indices of the posterior draws to be used. If `NULL` (the default) all draws are used in their native order.

response

Should results be returned on the response or logit scale? Defaults to `TRUE`, i.e. response scale. However, the autologistic parameter is not interpretable as a probability and is always returned on the logit scale regardless of the value of `response`

re_formula

formula containing group-level effects to be considered in the prediction. If `NULL` (default), include all group-level effects; if NA, include no group-level effects.

allow_new_levels

allow new levels for random effect terms in `new_data`? Will error if set to `FALSE` and new levels are provided in `new_data`.

sample_new_levels

If new_data is provided and contains random effect levels not present in the original data, how should predictions be handled? Passed directly to `brms::prepare_predictions`, which see.

Value

A list of sets of expected values (one per component). If `new_data` is a dataframe, each element contains one row per row of `new_data`. Otherwise, returns in the shape of the observation matrix/array used to format the flocker_data (but see `unit_level` parameter for further details).

Details

The probabilities returned are conditional probabilities (e.g. detection conditional on occupancy, colonization conditional on previous non-occupancy, etc). These probabilities are not conditioned on the observed histories (e.g. the occupancy probability is not fixed to one at sites with a detection; it is estimated only based on the covariates).

Examples

fitted_flocker(
  example_flocker_model_single, 
  summarise = TRUE, 
  draw_ids = 1:3
)
#> $linpred_occ
#> , , mean
#> 
#>             visit_1    visit_2    visit_3    visit_4
#> site_1   0.58755119 0.58755119 0.58755119 0.58755119
#> site_2   0.42755464 0.42755464 0.42755464 0.42755464
#> site_3   0.46994243 0.46994243 0.46994243 0.46994243
#> site_4   0.55659274 0.55659274 0.55659274 0.55659274
#> site_5   0.40599568 0.40599568 0.40599568 0.40599568
#> site_6   0.45216923 0.45216923 0.45216923 0.45216923
#> site_7   0.50053075 0.50053075 0.50053075 0.50053075
#> site_8   0.42437142 0.42437142 0.42437142 0.42437142
#> site_9   0.42530794 0.42530794 0.42530794 0.42530794
#> site_10  0.42846609 0.42846609 0.42846609 0.42846609
#> site_11  0.43613289 0.43613289 0.43613289 0.43613289
#> site_12  0.44423064 0.44423064 0.44423064 0.44423064
#> site_13  0.48449359 0.48449359 0.48449359 0.48449359
#> site_14  0.50128241 0.50128241 0.50128241 0.50128241
#> site_15  0.50642301 0.50642301 0.50642301 0.50642301
#> site_16  0.52788764 0.52788764 0.52788764 0.52788764
#> site_17  0.49451923 0.49451923 0.49451923 0.49451923
#> site_18  0.56429020 0.56429020 0.56429020 0.56429020
#> site_19  0.37135645 0.37135645 0.37135645 0.37135645
#> site_20  0.40819313 0.40819313 0.40819313 0.40819313
#> site_21  0.43028729 0.43028729 0.43028729 0.43028729
#> site_22  0.64952906 0.64952906 0.64952906 0.64952906
#> site_23  0.53902180 0.53902180 0.53902180 0.53902180
#> site_24  0.42324303 0.42324303 0.42324303 0.42324303
#> site_25  0.67094853 0.67094853 0.67094853 0.67094853
#> site_26  0.59286713 0.59286713 0.59286713 0.59286713
#> site_27  0.46239753 0.46239753 0.46239753 0.46239753
#> site_28  0.65433889 0.65433889 0.65433889 0.65433889
#> site_29  0.65298694 0.65298694 0.65298694 0.65298694
#> site_30  0.64803943 0.64803943 0.64803943 0.64803943
#> site_31  0.63352902 0.63352902 0.63352902 0.63352902
#> site_32  0.61453922 0.61453922 0.61453922 0.61453922
#> site_33  0.49766427 0.49766427 0.49766427 0.49766427
#> site_34  0.46106048 0.46106048 0.46106048 0.46106048
#> site_35  0.45267221 0.45267221 0.45267221 0.45267221
#> site_36  0.43042856 0.43042856 0.43042856 0.43042856
#> site_37  0.47412189 0.47412189 0.47412189 0.47412189
#> site_38  0.42397658 0.42397658 0.42397658 0.42397658
#> site_39  0.66459289 0.66459289 0.66459289 0.66459289
#> site_40  0.66987633 0.66987633 0.66987633 0.66987633
#> site_41  0.91158559 0.91158559 0.91158559 0.91158559
#> site_42  0.16472937 0.16472937 0.16472937 0.16472937
#> site_43  0.39640489 0.39640489 0.39640489 0.39640489
#> site_44  0.82663343 0.82663343 0.82663343 0.82663343
#> site_45  0.08574527 0.08574527 0.08574527 0.08574527
#> site_46  0.29050047 0.29050047 0.29050047 0.29050047
#> site_47  0.57811948 0.57811948 0.57811948 0.57811948
#> site_48  0.15109909 0.15109909 0.15109909 0.15109909
#> site_49  0.15503946 0.15503946 0.15503946 0.15503946
#> site_50  0.16875502 0.16875502 0.16875502 0.16875502
#> site_51  0.20470716 0.20470716 0.20470716 0.20470716
#> site_52  0.24643533 0.24643533 0.24643533 0.24643533
#> site_53  0.48494308 0.48494308 0.48494308 0.48494308
#> site_54  0.58229458 0.58229458 0.58229458 0.58229458
#> site_55  0.61029617 0.61029617 0.61029617 0.61029617
#> site_56  0.71571361 0.71571361 0.71571361 0.71571361
#> site_57  0.54404542 0.54404542 0.54404542 0.54404542
#> site_58  0.85089163 0.85089163 0.85089163 0.85089163
#> site_59  0.01854544 0.01854544 0.01854544 0.01854544
#> site_60  0.09236920 0.09236920 0.09236920 0.09236920
#> site_61  0.80682953 0.80682953 0.80682953 0.80682953
#> site_62  0.32041198 0.32041198 0.32041198 0.32041198
#> site_63  0.46448733 0.46448733 0.46448733 0.46448733
#> site_64  0.72884097 0.72884097 0.72884097 0.72884097
#> site_65  0.24748293 0.24748293 0.24748293 0.24748293
#> site_66  0.40438879 0.40438879 0.40438879 0.40438879
#> site_67  0.56456375 0.56456375 0.56456375 0.56456375
#> site_68  0.30957553 0.30957553 0.30957553 0.30957553
#> site_69  0.31276196 0.31276196 0.31276196 0.31276196
#> site_70  0.32351763 0.32351763 0.32351763 0.32351763
#> site_71  0.34967513 0.34967513 0.34967513 0.34967513
#> site_72  0.37732506 0.37732506 0.37732506 0.37732506
#> site_73  0.51277479 0.51277479 0.51277479 0.51277479
#> site_74  0.56694726 0.56694726 0.56694726 0.56694726
#> site_75  0.58313251 0.58313251 0.58313251 0.58313251
#> site_76  0.64833383 0.64833383 0.64833383 0.64833383
#> site_77  0.54535328 0.54535328 0.54535328 0.54535328
#> site_78  0.74906901 0.74906901 0.74906901 0.74906901
#> site_79  0.13395025 0.13395025 0.13395025 0.13395025
#> site_80  0.25485496 0.25485496 0.25485496 0.25485496
#> site_81  0.73475494 0.73475494 0.73475494 0.73475494
#> site_82  0.39391503 0.39391503 0.39391503 0.39391503
#> site_83  0.48932387 0.48932387 0.48932387 0.48932387
#> site_84  0.67297700 0.67297700 0.67297700 0.67297700
#> site_85  0.34448893 0.34448893 0.34448893 0.34448893
#> site_86  0.44962458 0.44962458 0.44962458 0.44962458
#> site_87  0.55623235 0.55623235 0.55623235 0.55623235
#> site_88  0.38665711 0.38665711 0.38665711 0.38665711
#> site_89  0.38879371 0.38879371 0.38879371 0.38879371
#> site_90  0.39599097 0.39599097 0.39599097 0.39599097
#> site_91  0.41341412 0.41341412 0.41341412 0.41341412
#> site_92  0.43174024 0.43174024 0.43174024 0.43174024
#> site_93  0.52140417 0.52140417 0.52140417 0.52140417
#> site_94  0.55784989 0.55784989 0.55784989 0.55784989
#> site_95  0.56887494 0.56887494 0.56887494 0.56887494
#> site_96  0.61418366 0.61418366 0.61418366 0.61418366
#> site_97  0.54324682 0.54324682 0.54324682 0.54324682
#> site_98  0.68844283 0.68844283 0.68844283 0.68844283
#> site_99  0.26209910 0.26209910 0.26209910 0.26209910
#> site_100 0.34956026 0.34956026 0.34956026 0.34956026
#> 
#> , , Q5
#> 
#>             visit_1    visit_2    visit_3    visit_4
#> site_1   0.26400531 0.26400531 0.26400531 0.26400531
#> site_2   0.29052972 0.29052972 0.29052972 0.29052972
#> site_3   0.40030785 0.40030785 0.40030785 0.40030785
#> site_4   0.30910630 0.30910630 0.30910630 0.30910630
#> site_5   0.22845274 0.22845274 0.22845274 0.22845274
#> site_6   0.35775006 0.35775006 0.35775006 0.35775006
#> site_7   0.39282597 0.39282597 0.39282597 0.39282597
#> site_8   0.28138206 0.28138206 0.28138206 0.28138206
#> site_9   0.28407386 0.28407386 0.28407386 0.28407386
#> site_10  0.29314813 0.29314813 0.29314813 0.29314813
#> site_11  0.31516043 0.31516043 0.31516043 0.31516043
#> site_12  0.33839090 0.33839090 0.33839090 0.33839090
#> site_13  0.41859788 0.41859788 0.41859788 0.41859788
#> site_14  0.39164582 0.39164582 0.39164582 0.39164582
#> site_15  0.38363348 0.38363348 0.38363348 0.38363348
#> site_16  0.35110985 0.35110985 0.35110985 0.35110985
#> site_17  0.40234890 0.40234890 0.40234890 0.40234890
#> site_18  0.29795515 0.29795515 0.29795515 0.29795515
#> site_19  0.12744177 0.12744177 0.12744177 0.12744177
#> site_20  0.23479623 0.23479623 0.23479623 0.23479623
#> site_21  0.06091115 0.06091115 0.06091115 0.06091115
#> site_22  0.44699126 0.44699126 0.44699126 0.44699126
#> site_23  0.45939381 0.45939381 0.45939381 0.45939381
#> site_24  0.07539719 0.07539719 0.07539719 0.07539719
#> site_25  0.42639194 0.42639194 0.42639194 0.42639194
#> site_26  0.46438015 0.46438015 0.46438015 0.46438015
#> site_27  0.23331778 0.23331778 0.23331778 0.23331778
#> site_28  0.44413459 0.44413459 0.44413459 0.44413459
#> site_29  0.44498062 0.44498062 0.44498062 0.44498062
#> site_30  0.44779959 0.44779959 0.44779959 0.44779959
#> site_31  0.45444522 0.45444522 0.45444522 0.45444522
#> site_32  0.46120495 0.46120495 0.46120495 0.46120495
#> site_33  0.34305495 0.34305495 0.34305495 0.34305495
#> site_34  0.22903009 0.22903009 0.22903009 0.22903009
#> site_35  0.20174721 0.20174721 0.20174721 0.20174721
#> site_36  0.12217960 0.12217960 0.12217960 0.12217960
#> site_37  0.27039800 0.27039800 0.27039800 0.27039800
#> site_38  0.06956330 0.06956330 0.06956330 0.06956330
#> site_39  0.38237258 0.38237258 0.38237258 0.38237258
#> site_40  0.42864690 0.42864690 0.42864690 0.42864690
#> site_41  0.81532586 0.81532586 0.81532586 0.81532586
#> site_42  0.11436834 0.11436834 0.11436834 0.11436834
#> site_43  0.25816491 0.25816491 0.25816491 0.25816491
#> site_44  0.67458290 0.67458290 0.67458290 0.67458290
#> site_45  0.06486401 0.06486401 0.06486401 0.06486401
#> site_46  0.19137713 0.19137713 0.19137713 0.19137713
#> site_47  0.39394267 0.39394267 0.39394267 0.39394267
#> site_48  0.10595685 0.10595685 0.10595685 0.10595685
#> site_49  0.10839103 0.10839103 0.10839103 0.10839103
#> site_50  0.11684883 0.11684883 0.11684883 0.11684883
#> site_51  0.13898529 0.13898529 0.13898529 0.13898529
#> site_52  0.16481477 0.16481477 0.16481477 0.16481477
#> site_53  0.31986711 0.31986711 0.31986711 0.31986711
#> site_54  0.39754437 0.39754437 0.39754437 0.39754437
#> site_55  0.42243708 0.42243708 0.42243708 0.42243708
#> site_56  0.52991491 0.52991491 0.52991491 0.52991491
#> site_57  0.36552560 0.36552560 0.36552560 0.36552560
#> site_58  0.71177864 0.71177864 0.71177864 0.71177864
#> site_59  0.01323198 0.01323198 0.01323198 0.01323198
#> site_60  0.06912891 0.06912891 0.06912891 0.06912891
#> site_61  0.69488124 0.69488124 0.69488124 0.69488124
#> site_62  0.25561546 0.25561546 0.25561546 0.25561546
#> site_63  0.42140239 0.42140239 0.42140239 0.42140239
#> site_64  0.62247101 0.62247101 0.62247101 0.62247101
#> site_65  0.17563861 0.17563861 0.17563861 0.17563861
#> site_66  0.35659404 0.35659404 0.35659404 0.35659404
#> site_67  0.49267637 0.49267637 0.49267637 0.49267637
#> site_68  0.24320985 0.24320985 0.24320985 0.24320985
#> site_69  0.24684055 0.24684055 0.24684055 0.24684055
#> site_70  0.25920076 0.25920076 0.25920076 0.25920076
#> site_71  0.28989536 0.28989536 0.28989536 0.28989536
#> site_72  0.32322036 0.32322036 0.32322036 0.32322036
#> site_73  0.45544300 0.45544300 0.45544300 0.45544300
#> site_74  0.49441592 0.49441592 0.49441592 0.49441592
#> site_75  0.50630253 0.50630253 0.50630253 0.50630253
#> site_76  0.55583592 0.55583592 0.55583592 0.55583592
#> site_77  0.47874797 0.47874797 0.47874797 0.47874797
#> site_78  0.63997410 0.63997410 0.63997410 0.63997410
#> site_79  0.06936262 0.06936262 0.06936262 0.06936262
#> site_80  0.18331803 0.18331803 0.18331803 0.18331803
#> site_81  0.65438009 0.65438009 0.65438009 0.65438009
#> site_82  0.28346440 0.28346440 0.28346440 0.28346440
#> site_83  0.40622657 0.40622657 0.40622657 0.40622657
#> site_84  0.58771548 0.58771548 0.58771548 0.58771548
#> site_85  0.20657964 0.20657964 0.20657964 0.20657964
#> site_86  0.36724128 0.36724128 0.36724128 0.36724128
#> site_87  0.47250143 0.47250143 0.47250143 0.47250143
#> site_88  0.27185965 0.27185965 0.27185965 0.27185965
#> site_89  0.27526632 0.27526632 0.27526632 0.27526632
#> site_90  0.28680009 0.28680009 0.28680009 0.28680009
#> site_91  0.31505671 0.31505671 0.31505671 0.31505671
#> site_92  0.34520797 0.34520797 0.34520797 0.34520797
#> site_93  0.43782052 0.43782052 0.43782052 0.43782052
#> site_94  0.47412660 0.47412660 0.47412660 0.47412660
#> site_95  0.48524608 0.48524608 0.48524608 0.48524608
#> site_96  0.53138976 0.53138976 0.53138976 0.53138976
#> site_97  0.45950626 0.45950626 0.45950626 0.45950626
#> site_98  0.60356231 0.60356231 0.60356231 0.60356231
#> site_99  0.09559974 0.09559974 0.09559974 0.09559974
#> site_100 0.21420572 0.21420572 0.21420572 0.21420572
#> 
#> , , Q95
#> 
#>             visit_1    visit_2    visit_3    visit_4
#> site_1   0.80034699 0.80034699 0.80034699 0.80034699
#> site_2   0.54255531 0.54255531 0.54255531 0.54255531
#> site_3   0.55458864 0.55458864 0.55458864 0.55458864
#> site_4   0.73724112 0.73724112 0.73724112 0.73724112
#> site_5   0.59126456 0.59126456 0.59126456 0.59126456
#> site_6   0.52031876 0.52031876 0.52031876 0.52031876
#> site_7   0.61915893 0.61915893 0.61915893 0.61915893
#> site_8   0.54938031 0.54938031 0.54938031 0.54938031
#> site_9   0.54736060 0.54736060 0.54736060 0.54736060
#> site_10  0.54062158 0.54062158 0.54062158 0.54062158
#> site_11  0.52469428 0.52469428 0.52469428 0.52469428
#> site_12  0.50847282 0.50847282 0.50847282 0.50847282
#> site_13  0.58318325 0.58318325 0.58318325 0.58318325
#> site_14  0.62081327 0.62081327 0.62081327 0.62081327
#> site_15  0.63205896 0.63205896 0.63205896 0.63205896
#> site_16  0.67790105 0.67790105 0.67790105 0.67790105
#> site_17  0.60583054 0.60583054 0.60583054 0.60583054
#> site_18  0.75294192 0.75294192 0.75294192 0.75294192
#> site_19  0.68741835 0.68741835 0.68741835 0.68741835
#> site_20  0.58600540 0.58600540 0.58600540 0.58600540
#> site_21  0.71034327 0.71034327 0.71034327 0.71034327
#> site_22  0.83438446 0.83438446 0.83438446 0.83438446
#> site_23  0.66750750 0.66750750 0.66750750 0.66750750
#> site_24  0.69753300 0.69753300 0.69753300 0.69753300
#> site_25  0.91921456 0.91921456 0.91921456 0.91921456
#> site_26  0.68045703 0.68045703 0.68045703 0.68045703
#> site_27  0.67735736 0.67735736 0.67735736 0.67735736
#> site_28  0.85110410 0.85110410 0.85110410 0.85110410
#> site_29  0.84634790 0.84634790 0.84634790 0.84634790
#> site_30  0.82930671 0.82930671 0.82930671 0.82930671
#> site_31  0.78145880 0.78145880 0.78145880 0.78145880
#> site_32  0.72138498 0.72138498 0.72138498 0.72138498
#> site_33  0.67168482 0.67168482 0.67168482 0.67168482
#> site_34  0.67762141 0.67762141 0.67762141 0.67762141
#> site_35  0.67942559 0.67942559 0.67942559 0.67942559
#> site_36  0.68698989 0.68698989 0.68698989 0.68698989
#> site_37  0.67524134 0.67524134 0.67524134 0.67524134
#> site_38  0.70052858 0.70052858 0.70052858 0.70052858
#> site_39  0.96045310 0.96045310 0.96045310 0.96045310
#> site_40  0.91335508 0.91335508 0.91335508 0.91335508
#> site_41  0.97841749 0.97841749 0.97841749 0.97841749
#> site_42  0.20139756 0.20139756 0.20139756 0.20139756
#> site_43  0.49281650 0.49281650 0.49281650 0.49281650
#> site_44  0.93577835 0.93577835 0.93577835 0.93577835
#> site_45  0.10757839 0.10757839 0.10757839 0.10757839
#> site_46  0.35298198 0.35298198 0.35298198 0.35298198
#> site_47  0.71439635 0.71439635 0.71439635 0.71439635
#> site_48  0.18550797 0.18550797 0.18550797 0.18550797
#> site_49  0.19011265 0.19011265 0.19011265 0.19011265
#> site_50  0.20607024 0.20607024 0.20607024 0.20607024
#> site_51  0.24743054 0.24743054 0.24743054 0.24743054
#> site_52  0.29471580 0.29471580 0.29471580 0.29471580
#> site_53  0.60485466 0.60485466 0.60485466 0.60485466
#> site_54  0.71904668 0.71904668 0.71904668 0.71904668
#> site_55  0.74957400 0.74957400 0.74957400 0.74957400
#> site_56  0.85257499 0.85257499 0.85257499 0.85257499
#> site_57  0.67555619 0.67555619 0.67555619 0.67555619
#> site_58  0.94993812 0.94993812 0.94993812 0.94993812
#> site_59  0.02397357 0.02397357 0.02397357 0.02397357
#> site_60  0.11562887 0.11562887 0.11562887 0.11562887
#> site_61  0.88009790 0.88009790 0.88009790 0.88009790
#> site_62  0.37986178 0.37986178 0.37986178 0.37986178
#> site_63  0.52875961 0.52875961 0.52875961 0.52875961
#> site_64  0.79143197 0.79143197 0.79143197 0.79143197
#> site_65  0.29827480 0.29827480 0.29827480 0.29827480
#> site_66  0.46770171 0.46770171 0.46770171 0.46770171
#> site_67  0.63119650 0.63119650 0.63119650 0.63119650
#> site_68  0.36807503 0.36807503 0.36807503 0.36807503
#> site_69  0.37155253 0.37155253 0.37155253 0.37155253
#> site_70  0.38321929 0.38321929 0.38321929 0.38321929
#> site_71  0.41114841 0.41114841 0.41114841 0.41114841
#> site_72  0.44002629 0.44002629 0.44002629 0.44002629
#> site_73  0.57874029 0.57874029 0.57874029 0.57874029
#> site_74  0.63358507 0.63358507 0.63358507 0.63358507
#> site_75  0.64974912 0.64974912 0.64974912 0.64974912
#> site_76  0.71395689 0.71395689 0.71395689 0.71395689
#> site_77  0.61186572 0.61186572 0.61186572 0.61186572
#> site_78  0.81560539 0.81560539 0.81560539 0.81560539
#> site_79  0.18212972 0.18212972 0.18212972 0.18212972
#> site_80  0.30677163 0.30677163 0.30677163 0.30677163
#> site_81  0.83647488 0.83647488 0.83647488 0.83647488
#> site_82  0.55924249 0.55924249 0.55924249 0.55924249
#> site_83  0.60260244 0.60260244 0.60260244 0.60260244
#> site_84  0.75314696 0.75314696 0.75314696 0.75314696
#> site_85  0.53572702 0.53572702 0.53572702 0.53572702
#> site_86  0.58359454 0.58359454 0.58359454 0.58359454
#> site_87  0.63380286 0.63380286 0.63380286 0.63380286
#> site_88  0.55599013 0.55599013 0.55599013 0.55599013
#> site_89  0.55695366 0.55695366 0.55695366 0.55695366
#> site_90  0.56016222 0.56016222 0.56016222 0.56016222
#> site_91  0.56771403 0.56771403 0.56771403 0.56771403
#> site_92  0.57537637 0.57537637 0.57537637 0.57537637
#> site_93  0.61764580 0.61764580 0.61764580 0.61764580
#> site_94  0.63455096 0.63455096 0.63455096 0.63455096
#> site_95  0.63964796 0.63964796 0.63964796 0.63964796
#> site_96  0.66482991 0.66482991 0.66482991 0.66482991
#> site_97  0.62779189 0.62779189 0.62779189 0.62779189
#> site_98  0.77509414 0.77509414 0.77509414 0.77509414
#> site_99  0.48505077 0.48505077 0.48505077 0.48505077
#> site_100 0.53830792 0.53830792 0.53830792 0.53830792
#> 
#> 
#> $linpred_det
#> , , mean
#> 
#>             visit_1   visit_2   visit_3    visit_4
#> site_1   0.14697785 0.5174280 0.3923756 0.17171877
#> site_2   0.27551585 0.6522097 0.4821800 0.25890827
#> site_3   0.54709452 0.7392291 0.2220486 0.56550390
#> site_4   0.34416333 0.2782618 0.1518496 0.22455404
#> site_5   0.55120264 0.6840826 0.5337575 0.72968974
#> site_6   0.38870062 0.6491010 0.2792436 0.25213053
#> site_7   0.78832736 0.2779319 0.3938057 0.45889106
#> site_8   0.44849146 0.6325001 0.6089222 0.50159906
#> site_9   0.55263966 0.5338612 0.7902826 0.39574413
#> site_10  0.33610687 0.3920462 0.5782934 0.48236854
#> site_11  0.43884997 0.3288622 0.4986144 0.35701007
#> site_12  0.21966548 0.4091210 0.6464682 0.37283013
#> site_13  0.49443922 0.1760459 0.3673286 0.47701197
#> site_14  0.39254167 0.3573833 0.2483286 0.22506425
#> site_15  0.20164626 0.3480261 0.2087502 0.25853001
#> site_16  0.25316147 0.6150712 0.2104020 0.32150749
#> site_17  0.36789032 0.2224537 0.3420926 0.62163061
#> site_18  0.27919360 0.2297035 0.1860774 0.09646986
#> site_19  0.80476202 0.4816874 0.7800270 0.84184058
#> site_20  0.31322132 0.3000433 0.4948070 0.32154265
#> site_21  0.40077934 0.6841599 0.6252437 0.44132183
#> site_22  0.39527071 0.6629290 0.5453181 0.37977345
#> site_23  0.63002285 0.7711876 0.3843854 0.64144423
#> site_24  0.57552493 0.5269551 0.3707978 0.47484163
#> site_25  0.56157400 0.6642034 0.5498969 0.70990479
#> site_26  0.51200168 0.6835591 0.4257199 0.39936962
#> site_27  0.83417597 0.4740463 0.5623011 0.60221017
#> site_28  0.51951555 0.6440115 0.6264497 0.55389094
#> site_29  0.58851441 0.5760762 0.8022141 0.48559856
#> site_30  0.44654666 0.4869705 0.6096347 0.54656043
#> site_31  0.52755852 0.4499484 0.5661009 0.47127794
#> site_32  0.35572506 0.5171552 0.6747869 0.49204990
#> site_33  0.61032432 0.3365055 0.5298647 0.59993837
#> site_34  0.56213179 0.5384230 0.4455635 0.41954328
#> site_35  0.39515744 0.5364023 0.4045309 0.46116491
#> site_36  0.47678166 0.7072795 0.4285074 0.53593338
#> site_37  0.53961365 0.4096141 0.5211406 0.69386443
#> site_38  0.53464158 0.4882922 0.4346807 0.25502359
#> site_39  0.75395793 0.4561892 0.7144967 0.82107672
#> site_40  0.40399234 0.3935612 0.5276517 0.41040274
#> site_41  0.49248905 0.1917149 0.2286948 0.43260483
#> site_42  0.44814824 0.1777094 0.2711409 0.46924723
#> site_43  0.20805971 0.1295154 0.4708548 0.20000997
#> site_44  0.26507655 0.3138235 0.5195214 0.37566996
#> site_45  0.25098706 0.1715380 0.2619646 0.14463491
#> site_46  0.30976051 0.1700087 0.4121090 0.44682151
#> site_47  0.11254264 0.3624593 0.2684230 0.23396720
#> site_48  0.29638825 0.1895304 0.2019661 0.26232032
#> site_49  0.23184229 0.2424537 0.1074083 0.33374701
#> site_50  0.38123457 0.3327751 0.2155264 0.27013487
#> site_51  0.29045966 0.3783513 0.2535368 0.35247491
#> site_52  0.50470203 0.3027846 0.1737624 0.33005369
#> site_53  0.22511371 0.5400732 0.2972225 0.23332330
#> site_54  0.26870285 0.2915773 0.3978818 0.43173287
#> site_55  0.46606524 0.2945238 0.4533223 0.37954635
#> site_56  0.36559703 0.1692831 0.4275916 0.29884080
#> site_57  0.28919547 0.4431151 0.3081182 0.17124814
#> site_58  0.30701391 0.3609073 0.4319488 0.68001682
#> site_59  0.10962118 0.3625737 0.1290891 0.08249990
#> site_60  0.43419807 0.4484578 0.2845565 0.42551559
#> site_61  0.32376941 0.4038029 0.3829150 0.33285113
#> site_62  0.37853694 0.4541005 0.4200487 0.37414215
#> site_63  0.42412049 0.4757677 0.3543278 0.42779302
#> site_64  0.37601062 0.3620797 0.3233188 0.34832392
#> site_65  0.43802820 0.4662188 0.4348966 0.47959539
#> site_66  0.39750400 0.4492711 0.3737289 0.36649083
#> site_67  0.49859993 0.3661429 0.3909422 0.40308708
#> site_68  0.41480272 0.4500528 0.4448843 0.42429530
#> site_69  0.43345964 0.4299476 0.5033378 0.40492526
#> site_70  0.39228617 0.4034120 0.4377537 0.41986728
#> site_71  0.41024984 0.3888128 0.4210897 0.39468756
#> site_72  0.35823222 0.4029969 0.4498597 0.39602646
#> site_73  0.41187607 0.3357445 0.3879049 0.40863782
#> site_74  0.39060650 0.3837052 0.3582794 0.35146352
#> site_75  0.34338058 0.3812004 0.3457860 0.36054419
#> site_76  0.35736975 0.4305753 0.3449403 0.37351873
#> site_77  0.38665627 0.3513265 0.3814031 0.43649641
#> site_78  0.36201966 0.3496900 0.3364395 0.29581379
#> site_79  0.51131263 0.4406236 0.5016822 0.52930270
#> site_80  0.39356805 0.3906003 0.4274164 0.39537908
#> site_81  0.53909209 0.1784563 0.2358909 0.47863907
#> site_82  0.52830813 0.2261382 0.3406023 0.54949857
#> site_83  0.25321098 0.1382770 0.5431714 0.24260910
#> site_84  0.29906892 0.3581649 0.5735617 0.42731169
#> site_85  0.32237916 0.2249549 0.3349922 0.18832068
#> site_86  0.37876597 0.2074329 0.4872343 0.52240806
#> site_87  0.09109821 0.4261203 0.3201262 0.27758233
#> site_88  0.36972967 0.2426569 0.2585552 0.33119412
#> site_89  0.29509615 0.3077594 0.1221226 0.41031467
#> site_90  0.45982196 0.4086496 0.2743723 0.33924707
#> site_91  0.36061708 0.4553890 0.3182340 0.42820077
#> site_92  0.58166157 0.3726794 0.2153463 0.40259885
#> site_93  0.27096831 0.6087522 0.3576710 0.28137445
#> site_94  0.32026537 0.3472250 0.4632629 0.49798094
#> site_95  0.53148490 0.3493776 0.5187477 0.44303700
#> site_96  0.42350783 0.1763846 0.4883747 0.34888445
#> site_97  0.34610162 0.5107907 0.3677533 0.19375240
#> site_98  0.34772395 0.4091309 0.4844255 0.72759508
#> site_99  0.15328275 0.4514358 0.1810522 0.11039857
#> site_100 0.51789079 0.5323197 0.3600002 0.50907590
#> 
#> , , Q5
#> 
#>             visit_1    visit_2    visit_3    visit_4
#> site_1   0.04010375 0.36198594 0.30930049 0.05550500
#> site_2   0.17647725 0.47262309 0.40600620 0.15381094
#> site_3   0.41672410 0.50777695 0.10698037 0.42330425
#> site_4   0.25896590 0.16570974 0.04271374 0.10311372
#> site_5   0.44073730 0.49710738 0.43452122 0.52394047
#> site_6   0.34351113 0.46257255 0.17872488 0.14374619
#> site_7   0.54206982 0.17178454 0.34132489 0.36901775
#> site_8   0.39285238 0.46502608 0.45518601 0.41582131
#> site_9   0.43349411 0.42683224 0.56571024 0.35900951
#> site_10  0.26389394 0.35247807 0.44174480 0.40569584
#> site_11  0.38374635 0.25181538 0.40952204 0.29520335
#> site_12  0.10419973 0.36774782 0.46404324 0.31910400
#> site_13  0.38974191 0.06023892 0.30484470 0.38225051
#> site_14  0.34053401 0.28700395 0.13459340 0.10842513
#> site_15  0.08470519 0.27183971 0.09150954 0.14642460
#> site_16  0.13775999 0.42493922 0.09153374 0.22923287
#> site_17  0.30440076 0.10619637 0.26429644 0.43753033
#> site_18  0.16589281 0.10766805 0.06764438 0.01449519
#> site_19  0.60465187 0.43540944 0.58110907 0.64867350
#> site_20  0.23242562 0.21317467 0.42010973 0.24483269
#> site_21  0.22102882 0.48207593 0.41317432 0.24895405
#> site_22  0.35718552 0.63871347 0.53648216 0.33829808
#> site_23  0.56420007 0.70493223 0.35466935 0.57481678
#> site_24  0.40357921 0.35935702 0.23974252 0.31602029
#> site_25  0.52564724 0.64238925 0.51067248 0.68479299
#> site_26  0.48558854 0.63690142 0.41325736 0.39034742
#> site_27  0.75498293 0.38734374 0.46283257 0.49937883
#> site_28  0.50705222 0.62286216 0.60759217 0.54388035
#> site_29  0.57468845 0.56389444 0.76870149 0.46583281
#> site_30  0.42227419 0.47312993 0.59197612 0.53767816
#> site_31  0.51978171 0.43948151 0.55259148 0.46639344
#> site_32  0.33378230 0.50190815 0.63915888 0.48073876
#> site_33  0.52722674 0.29820996 0.45547164 0.51765018
#> site_34  0.46171270 0.44076442 0.36328341 0.34262498
#> site_35  0.31727664 0.43234162 0.32450509 0.36924956
#> site_36  0.35448673 0.57612992 0.31643020 0.40437050
#> site_37  0.45062800 0.34332452 0.43467966 0.59718995
#> site_38  0.35587259 0.31664859 0.27526119 0.15861480
#> site_39  0.70566101 0.30858814 0.64997106 0.79197526
#> site_40  0.33353846 0.32116527 0.48672229 0.34119568
#> site_41  0.48107751 0.04371267 0.08391274 0.38886440
#> site_42  0.32336682 0.08668016 0.20399464 0.32970971
#> site_43  0.10797369 0.02719109 0.37319083 0.09816461
#> site_44  0.14741731 0.21795775 0.47458716 0.31446815
#> site_45  0.18194313 0.08736482 0.19638825 0.05866690
#> site_46  0.25208876 0.06994950 0.33779587 0.34833800
#> site_47  0.01017934 0.31753659 0.17686901 0.12968643
#> site_48  0.23949152 0.10163551 0.11685216 0.19290252
#> site_49  0.15374296 0.16698012 0.02142074 0.28698323
#> site_50  0.30440099 0.28999447 0.13257688 0.20244435
#> site_51  0.22854551 0.31150399 0.17892494 0.30375801
#> site_52  0.35803529 0.24400106 0.07661407 0.28329358
#> site_53  0.12423476 0.40795063 0.22551226 0.13496143
#> site_54  0.17696034 0.21015735 0.37120049 0.39136978
#> site_55  0.40710366 0.21235767 0.40311773 0.34178429
#> site_56  0.31269237 0.04527395 0.40978952 0.20939225
#> site_57  0.20944720 0.38836990 0.23777652 0.05786256
#> site_58  0.20368548 0.28671641 0.40203638 0.52878153
#> site_59  0.03864120 0.22638815 0.05621379 0.01842753
#> site_60  0.29746446 0.30166313 0.22662869 0.29491647
#> site_61  0.17000293 0.28499999 0.26894829 0.18403542
#> site_62  0.31557086 0.36569203 0.35310150 0.31166150
#> site_63  0.35371285 0.34534749 0.34197625 0.35310680
#> site_64  0.29455267 0.26947853 0.20303224 0.24528669
#> site_65  0.33063198 0.35789309 0.32773624 0.37146342
#> site_66  0.36523984 0.35620791 0.34933705 0.34248509
#> site_67  0.33170403 0.33575457 0.34244770 0.34168520
#> site_68  0.34283597 0.36793305 0.36890436 0.35156450
#> site_69  0.36174745 0.35844988 0.35745157 0.33546544
#> site_70  0.32935914 0.33940766 0.36829781 0.35448170
#> site_71  0.35832356 0.33863474 0.36785588 0.34400431
#> site_72  0.32218352 0.36456079 0.35926094 0.35807051
#> site_73  0.35050583 0.29285697 0.35269141 0.35103742
#> site_74  0.34200492 0.34248358 0.32018695 0.30741311
#> site_75  0.28742622 0.33949558 0.29186719 0.31932167
#> site_76  0.29173884 0.32342908 0.26924563 0.32144429
#> site_77  0.34647377 0.31372716 0.34684891 0.34315232
#> site_78  0.26059778 0.23909077 0.21660756 0.15273190
#> site_79  0.32805088 0.25401028 0.31593864 0.35298896
#> site_80  0.29540760 0.29288295 0.32511961 0.29695348
#> site_81  0.49548685 0.07544114 0.12753842 0.44977508
#> site_82  0.51807570 0.12314833 0.26443812 0.53441795
#> site_83  0.15213616 0.04721567 0.52353928 0.14022991
#> site_84  0.20216342 0.27936941 0.52774724 0.37760635
#> site_85  0.24059915 0.12221466 0.25772568 0.08619557
#> site_86  0.31700786 0.10385481 0.47785583 0.51355667
#> site_87  0.02204529 0.38184748 0.23390055 0.18005755
#> site_88  0.30542057 0.14115025 0.15933542 0.25171015
#> site_89  0.20406592 0.22042169 0.03695351 0.36429357
#> site_90  0.43807807 0.36166472 0.17811883 0.26253500
#> site_91  0.29193764 0.43094980 0.23382137 0.39020697
#> site_92  0.55321637 0.30866954 0.11178881 0.35195068
#> site_93  0.17253375 0.56385497 0.28572159 0.18512253
#> site_94  0.23401629 0.26988809 0.43672725 0.48627406
#> site_95  0.50922400 0.27233738 0.50006724 0.40615094
#> site_96  0.37519778 0.07536984 0.47150014 0.26964156
#> site_97  0.26898686 0.49690746 0.29878178 0.09061012
#> site_98  0.26434985 0.35027440 0.46147554 0.63756208
#> site_99  0.05698943 0.42851824 0.07967770 0.02966846
#> site_100 0.50333473 0.51443040 0.29244563 0.49652867
#> 
#> , , Q95
#> 
#>            visit_1   visit_2   visit_3   visit_4
#> site_1   0.2126811 0.6405930 0.5176985 0.2486647
#> site_2   0.3354290 0.8006888 0.5355237 0.3240346
#> site_3   0.6281698 0.9184819 0.2921811 0.6597567
#> site_4   0.4558494 0.3857742 0.2247307 0.3178115
#> site_5   0.6373846 0.8462964 0.6075804 0.9040955
#> site_6   0.4557441 0.7962458 0.3492817 0.3183929
#> site_7   0.9658983 0.3651148 0.4800045 0.5453129
#> site_8   0.5007419 0.7704084 0.7328933 0.5528997
#> site_9   0.6390913 0.6069068 0.9625354 0.4498129
#> site_10  0.3973457 0.4481113 0.6823586 0.5361785
#> site_11  0.4971920 0.3938015 0.5559416 0.4203214
#> site_12  0.2981258 0.4713505 0.7921306 0.4382408
#> site_13  0.5733889 0.2597737 0.4495462 0.5563937
#> site_14  0.4790097 0.4467646 0.3304151 0.3006563
#> site_15  0.2708650 0.4398651 0.2799889 0.3445364
#> site_16  0.3452510 0.7424797 0.2887581 0.4218089
#> site_17  0.4539192 0.2950966 0.4296819 0.7531659
#> site_18  0.3899248 0.3276070 0.2639414 0.1738668
#> site_19  0.9674461 0.5199065 0.9501090 0.9852908
#> site_20  0.3654460 0.3522956 0.5401204 0.3735609
#> site_21  0.5909124 0.8893489 0.8447630 0.6456205
#> site_22  0.4247635 0.6935590 0.5533556 0.4112198
#> site_23  0.7123863 0.8589135 0.4072760 0.7253616
#> site_24  0.7597454 0.7060185 0.5041523 0.6430975
#> site_25  0.6040990 0.6893648 0.5945145 0.7285190
#> site_26  0.5436729 0.7502185 0.4342722 0.4079818
#> site_27  0.9286578 0.5655694 0.6738593 0.7196453
#> site_28  0.5359442 0.6659187 0.6441767 0.5656092
#> site_29  0.5981739 0.5838639 0.8498349 0.5057979
#> site_30  0.4673884 0.5016401 0.6293688 0.5531357
#> site_31  0.5379747 0.4578904 0.5864136 0.4759051
#> site_32  0.3713137 0.5376485 0.7298776 0.5059373
#> site_33  0.7086306 0.3644706 0.6131772 0.6967396
#> site_34  0.6746438 0.6464391 0.5300268 0.4960922
#> site_35  0.4707721 0.6508569 0.4832193 0.5571781
#> site_36  0.6062316 0.8531345 0.5438066 0.6785541
#> site_37  0.6387936 0.4738823 0.6162538 0.8090015
#> site_38  0.7251919 0.6705911 0.6022405 0.3447213
#> site_39  0.8194086 0.5975208 0.7930574 0.8648285
#> site_40  0.4657967 0.4564351 0.5719068 0.4715111
#> site_41  0.5061085 0.4004752 0.4216254 0.4896947
#> site_42  0.5451985 0.2526440 0.3329224 0.5699219
#> site_43  0.3014940 0.2601900 0.5418658 0.2981935
#> site_44  0.4021034 0.4187981 0.5520383 0.4370361
#> site_45  0.3214695 0.2299404 0.3353631 0.2182601
#> site_46  0.3648706 0.2701773 0.4846810 0.5258050
#> site_47  0.2638135 0.3983495 0.3500040 0.3377164
#> site_48  0.3652581 0.2545277 0.2592340 0.3239106
#> site_49  0.2840805 0.2980182 0.2126580 0.4083128
#> site_50  0.4647095 0.4046699 0.2679235 0.3311362
#> site_51  0.3512425 0.4554762 0.3058549 0.4236521
#> site_52  0.5991308 0.3611619 0.2653396 0.3923743
#> site_53  0.3205253 0.6217240 0.3447977 0.3236193
#> site_54  0.3507685 0.3581625 0.4356836 0.4748560
#> site_55  0.5118606 0.3637493 0.4968857 0.4136506
#> site_56  0.4047247 0.3299744 0.4526581 0.3850988
#> site_57  0.3513289 0.4928288 0.3570813 0.3045141
#> site_58  0.4250298 0.4414372 0.4607604 0.8038735
#> site_59  0.1645349 0.4959060 0.1736158 0.1489385
#> site_60  0.5442747 0.5610428 0.3615780 0.5340182
#> site_61  0.4863024 0.5881706 0.5627073 0.4995137
#> site_62  0.4346337 0.5888021 0.5189322 0.4256112
#> site_63  0.4883365 0.5954249 0.3662637 0.4960087
#> site_64  0.5033531 0.4875305 0.4390827 0.4711592
#> site_65  0.5816976 0.6375070 0.5753747 0.6632203
#> site_66  0.4472278 0.5562621 0.3991448 0.3847278
#> site_67  0.6162904 0.4110382 0.4415041 0.4566588
#> site_68  0.5121029 0.5833146 0.5724111 0.5312086
#> site_69  0.5483944 0.5413976 0.6916750 0.4910135
#> site_70  0.4617438 0.4843411 0.5534958 0.5175025
#> site_71  0.4894363 0.4462958 0.5111482 0.4581577
#> site_72  0.3822940 0.4659429 0.5645791 0.4520244
#> site_73  0.4505559 0.3645082 0.4154551 0.4438186
#> site_74  0.4421246 0.4334290 0.4038059 0.3965168
#> site_75  0.3946925 0.4373287 0.3973556 0.4133661
#> site_76  0.4400256 0.5291945 0.4258748 0.4577238
#> site_77  0.4278079 0.3870910 0.4212087 0.4932187
#> site_78  0.4994060 0.4845391 0.4677324 0.4093322
#> site_79  0.7759017 0.6371545 0.7587344 0.8060092
#> site_80  0.4874895 0.4812723 0.5573786 0.4912772
#> site_81  0.5775454 0.2433036 0.3100751 0.5259115
#> site_82  0.5424279 0.3059587 0.3985426 0.5740664
#> site_83  0.3224758 0.2076597 0.5620700 0.3133170
#> site_84  0.3657127 0.4188940 0.6037281 0.4758900
#> site_85  0.3900547 0.3093406 0.3993518 0.2733513
#> site_86  0.4204300 0.2839504 0.4940810 0.5315719
#> site_87  0.1449838 0.4597471 0.3712399 0.3376247
#> site_88  0.4197683 0.3214992 0.3352777 0.3923791
#> site_89  0.3648709 0.3746634 0.1939832 0.4469777
#> site_90  0.4784623 0.4452018 0.3476392 0.3973747
#> site_91  0.4109936 0.4740297 0.3803833 0.4564580
#> site_92  0.6210607 0.4177675 0.2928476 0.4379980
#> site_93  0.3348917 0.6590963 0.4001009 0.3432964
#> site_94  0.3715445 0.3951602 0.4889643 0.5159618
#> site_95  0.5462782 0.3982822 0.5353005 0.4745897
#> site_96  0.4648721 0.2416854 0.5153723 0.4032432
#> site_97  0.3925234 0.5251402 0.4108593 0.2640474
#> site_98  0.4120423 0.4635972 0.5227139 0.8194852
#> site_99  0.2433157 0.4893626 0.2748022 0.1884156
#> site_100 0.5294834 0.5510877 0.4166901 0.5166258
#> 
#>