Get expected values of the posterior predictive distribution for the occupancy component, the detection component, or both. Note that in the case of the detection component, you are conditioning on occupancy being 1. Note that these are values for a hypothetical set of unit covariates rather than the ones actually observed (some of which are known to be occupied, when Q == 1).

Get expected values of the posterior predictive distribution for the occupancy component, the detection component, or both. Note that in the case of the detection component, you are conditioning on occupancy being 1. Note that these are values for a hypothetical set of unit covariates rather than the ones actually observed (some of which are known to be occupied, when Q == 1).

fitted_flocker(
  flocker_fit,
  type,
  new_data = NULL,
  CI = c(0.05, 0.95),
  ndraws = NULL,
  response = TRUE,
  re_formula = NULL,
  summarise = TRUE,
  allow_new_levels = FALSE,
  sample_new_levels = "uncertainty"
)

Arguments

flocker_fit

A flocker_fit object.

type

Get posterior probabilities for 'occupancy', 'detection', or for 'both'.

new_data

Optional new data at which to evaluate occupancy predictions. If `NULL` (the default) expected values are generated for the original data.

CI

A vector of length 2 specifying the upper and lower bounds of the credible interval. Defaults to c(.05, .95)

ndraws

Positive integer indicating how many poserior draws should be used. If `NULL` (the default) all draws are used.

response

Should expected values be returned on the response or logit-scale? Defaults to `TRUE`, i.e. response scale.

re_formula

formula containing group-level effects to be considered in the prediction. If `NULL` (default), include all group-level effects; if NA, include no group-level effects.

summarise

if TRUE, return the expected value and upper and lower bound of the credible interval, otherwise return the posterior matrix.

allow_new_levels

allow new levels in new_data?

sample_new_levels

If new_data is provided and contains random effect levels not present in the original data, how should predictions be handled? Passed directly to brms::prepare_predictions.

Value

A data.frame containing the columns used to generate expected values and the estimate and its credible interval (lower and upper bounds)